requirements.txt
asgiref==3.9.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
